+++ CHANGELOG.rst
@@ -1,12 +1,16 @@
 CHANGELOG
 =========
 
+Changelog has been moved to `http://docs.waldur.com/Changelog
+<http://docs.waldur.com/Changelog>`_
+
 Release NEXT
 ------------
 - Raise an ElasticsearchClientError if ELASTICSEARCH configuration keys are missing or empty.
 - Allow to filter user by civil number.
 - Don't render superuser status. Drop unused viewsets.
 - Add LDAP scheme to service settings backend_url validator.
+- Add organization cost limit.
 
 Release 0.135.0
 ---------------